requirements.txt
apns2==0.7.1
asgiref==3.2.10
astroid==2.4.2
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.1.1
Django==2.2.16
djangorestframework==3.11.1
drf-yasg==1.17.1
fcm-django==0.3.4
h2==2.6.2
hpack==3.0.0
http-ece==1.1.0
hyper==0.7.0
hyperframe==3.2.0
idna==2.10
inflection==0.5.1
isort==5.5.2
itypes==1.2.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
mysqlclient==2.0.1
packaging==20.4
Pillow==7.2.0
py-vapid==1.7.1
pycparser==2.20
pyfcm==1.4.7
PyJWT==1.7.1
pylint==2.6.0
pyparsing==2.4.7
pytz==2020.1
pywebpush==1.11.0
requests==2.24.0
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
typed-ast==1.4.1
uritemplate==3.0.1
urllib3==1.25.10
wrapt==1.12.1
